refactor: create shared basic blocks
